Added new client
