--- About and Contact/contactUs.css
@@ -1,23 +1,0 @@
-
-h1{
-    text-align: center;
-	color: #edf0f1;
-	padding-bottom:60px;
-	font-size:60px;
-    padding-top: 30px;
-	font-family: 'Poppins', sans-serif;
-}	
-
-
-body{
-	color:#edf0f1;
-	background-repeat:no-repeat;
-	background-attachment:fixed;
-	background-size:cover;
-	font-family: 'Poppins', sans-serif;
-	font-size:20px;
-}
-
-.info{
-	    margin: 12% 0 0 25%;
-}
