+++ client/.gitignore
@@ -1,5 +1,6 @@
-# Created by https://www.toptal.com/developers/gitignore/api/node,visualstudiocode
-# Edit at https://www.toptal.com/developers/gitignore?templates=node,visualstudiocode
+
+# Created by https://www.toptal.com/developers/gitignore/api/react,node,visualstudiocode
+# Edit at https://www.toptal.com/developers/gitignore?templates=react,node,visualstudiocode
 
 ### Node ###
 # Logs
@@ -141,6 +142,19 @@ dist
 # SvelteKit build / generate output
 .svelte-kit
 
+### react ###
+.DS_*
+**/*.backup.*
+**/*.back.*
+
+node_modules
+
+*.sublime*
+
+psd
+thumb
+sketch
+
 ### VisualStudioCode ###
 .vscode/*
 !.vscode/settings.json
@@ -162,4 +176,4 @@ dist
 
 # Support for Project snippet scope
 
-# End of https://www.toptal.com/developers/gitignore/api/node,visualstudiocode
+# End of https://www.toptal.com/developers/gitignore/api/react,node,visualstudiocode
